Tested the GO script. It works just fine.

+++ Build-Input-Files-for-Gene-Ontology/Main.sh
@@ -4,51 +4,76 @@
 # Date: 2012-10-10
 
 echo "Hi!"
+echo "Welcome to Ray Technologies, where everything runs in parallel"
+echo ""
 
-echo "This script will download these:"
-echo "- gene_association.goa_uniprot (uniprot -> gene ontology)"
-echo "- idmapping.dat  (uniprot -> EMBL_CDS)"
+echo "This script will download files and tell you what to do with them."
+echo ""
 
+echo "Computing nothing during 5 seconds, please wait."
 sleep 5
 
 echo ""
 
-echo "Downloading Uniprot-GOA gene_association.goa_uniprot.gz"
-wget ftp://ftp.ebi.ac.uk/pub/databases/GO/goa/UNIPROT/gene_association.goa_uniprot.gz
-gunzip gene_association.goa_uniprot.gz
+if test ! -f gene_association.goa_uniprot
+then
+	echo "Downloading Uniprot-GOA gene_association.goa_uniprot.gz"
+	wget ftp://ftp.ebi.ac.uk/pub/databases/GO/goa/UNIPROT/gene_association.goa_uniprot.gz
+	gunzip gene_association.goa_uniprot.gz
+fi
 
-echo "Downloading mappings idmapping.dat.gz"
-wget ftp://ftp.ebi.ac.uk/pub/databases/uniprot/current_release/knowledgebase/idmapping/idmapping.dat.gz
-gunzip idmapping.dat.gz
+if test ! -f idmapping.dat
+then
+	echo "Downloading mappings idmapping.dat.gz"
+	wget ftp://ftp.ebi.ac.uk/pub/databases/uniprot/current_release/knowledgebase/idmapping/idmapping.dat.gz
+	gunzip idmapping.dat.gz
+fi
 
-echo "Building annotations 000-Annotations.txt"
-./CreateBuild.py
+if test ! -f 000-Annotations.txt
+then
+	echo "Building annotations 000-Annotations.txt"
+	./CreateBuild.py
 
-ln -s Associations.txt 000-Annotations.txt
+	ln -s Associations.txt 000-Annotations.txt
+fi
 
-echo "Downloading cds.fasta.gz"
-wget ftp://ftp.ebi.ac.uk/pub/databases/embl/cds/cds.fasta.gz
-gunzip cds.fasta.gz 
+if test ! -f cds.fasta
+then
+	echo "Downloading cds.fasta.gz"
+	wget ftp://ftp.ebi.ac.uk/pub/databases/embl/cds/cds.fasta.gz
+	gunzip cds.fasta.gz 
+fi
 
-echo "Rebuilding the fasta file, removing unannotated entries..."
-./Rebuild-Fasta.py
-ln -s Sequences-With-Annotations.fasta 000-Sequences.fasta
+if test ! -f 000-Sequences.fasta
+then
+	echo "Rebuilding the fasta file, removing unannotated entries..."
+	./Rebuild-Fasta.py
+	ln -s Sequences-With-Annotations.fasta 000-Sequences.fasta
+fi
 
-echo "Downloading Gene Ontology file obo_format_1_2/gene_ontology_ext.obo"
-wget http://geneontology.org/ontology/obo_format_1_2/gene_ontology_ext.obo
+if test ! -f 000-Ontologies.txt
+then
+	echo "Downloading Gene Ontology file obo_format_1_2/gene_ontology_ext.obo"
+	wget http://geneontology.org/ontology/obo_format_1_2/gene_ontology_ext.obo
 
-ln -s gene_ontology_ext.obo 000-Ontologies.txt
+	ln -s gene_ontology_ext.obo 000-Ontologies.txt
+fi
 
+if test ! -d EMBL_CDS_Sequences
+then
+	echo "Generating parts."
 
-echo "Generating parts."
+	./SplitLargeFastaFileInSmallerParts.py 000-Sequences.fasta 128
 
-./SplitLargeFastaFileInSmallerParts.py 000-Sequences.fasta 128
+	mkdir EMBL_CDS_Sequences
+	mv 000-Sequences.Part.*.fasta EMBL_CDS_Sequences
+fi
 
-mkdir EMBL_CDS_Sequences
-mv 000-Sequences.fasta.Part.*.fasta EMBL_CDS_Sequences
-
-ln -s 000-Ontologies.txt OntologyTerms.txt
-ln -s 000-Annotations.txt Annotations.txt
+if test ! -f OntologyTerms.txt
+then
+	ln -s 000-Ontologies.txt OntologyTerms.txt
+	ln -s 000-Annotations.txt Annotations.txt
+fi
 
 echo "Thank you for your kind patience !"
 
@@ -57,6 +82,9 @@ echo ""
 echo "To use these builds with Ray Ontologies (via the Ray binary), add these options:"
 echo ""
 
-echo "-search $(pwd)/EMBL_CDS_Sequences \\" 
-echo "-gene-ontology $(pwd)/OntologyTerms.txt  $(pwd)/Annotations.txt \\"
+echo "-search \\"
+echo "    $(pwd)/EMBL_CDS_Sequences \\" 
+echo "-gene-ontology \\"
+echo "    $(pwd)/OntologyTerms.txt \\"
+echo "    $(pwd)/Annotations.txt \\"
